requirements.txt
torch
numpy
einops
tqdm
argbind
torchaudio
soundfile
click
PyYAML
accelerate
torchmetrics
munch
librosa==0.9.0
pandas
descript-audio-codec
transformers